add Susbcription Flow and update sub to backend in API subs history page resposne change handled

+++ app/src/main/java/com/jetsynthesys/rightlife/ui/settings/pojo/BundleShare.kt
@@ -6,9 +6,9 @@ import com.google.gson.annotations.SerializedName
 class BundleShare {
     @SerializedName("INR")
     @Expose
-    var inr: Int? = null
+    var inr: Double? = null
 
     @SerializedName("USD")
     @Expose
-    var usd: Int? = null
+    var usd: Double? = null
 }
